Database Functionality Addition
Added Database Functionality to the game. Now, users are able to register an account to keep track of score, wins, and losses. Furthermore, a basic main screen has been created. In the next iteration, I hope to create a class to manage this database, as all attempts to create this class during this iteration have failed, with more research and testing to be done.